Fixes bug in nuchal fold typing
The observed nuchal fold was declared as an int, but should be a double.

+++ down-syndrome-risk-assessment/RiskAssessmentModel.cs
@@ -317,8 +317,8 @@
             }
         }
 
-        private int _nuchalFoldObserved;
-        public int nuchalFoldObserved
+        private double _nuchalFoldObserved;
+        public double nuchalFoldObserved
         {
             get { return _nuchalFoldObserved; }
             set
